backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
SQLAlchemy
PyMySQL
python-dotenv
flask-jwt-extended
pydantic
